requirements.txt
selenium
pytesseract
pillow
opencv-python
numpy
pyautogui
termcolor
pandas
undetected-chromedriver
